new LongBuffers for TADs

+++ nd4j-backends/nd4j-backend-impls/nd4j-native/src/main/java/org/nd4j/linalg/cpu/nativecpu/CpuTADManager.java
@@ -8,6 +8,7 @@ import org.bytedeco.javacpp.Pointer;
 import org.nd4j.linalg.api.buffer.DataBuffer;
 import org.nd4j.linalg.api.buffer.DoubleBuffer;
 import org.nd4j.linalg.api.buffer.IntBuffer;
+import org.nd4j.linalg.api.buffer.LongBuffer;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.cache.ConstantHandler;
 import org.nd4j.linalg.cache.TADManager;
@@ -67,7 +68,7 @@ public class CpuTADManager implements TADManager {
                 offsetLength = array.lengthLong() / tadLength;
 
                 DataBuffer outputBuffer = new IntBuffer(targetRank * 2 + 4);
-                DataBuffer offsetsBuffer = new DoubleBuffer(offsetLength);
+                DataBuffer offsetsBuffer = new LongBuffer(offsetLength);
 
                 DataBuffer dimensionBuffer = constantHandler.getConstantBuffer(dimension);
                 Pointer dimensionPointer = dimensionBuffer.addressPointer();
